Fix PHP 8.4 SecurityBundle grind: DI, strict typing, and harness updates.
Align AuthenticatorManager wiring with the current constructor, coerce
request and time values under PHP 8.4 strict typing, and fix ExpressionLanguage,
rate limiting, login link, and OIDC JWT handling used by SecurityBundle tests.

Add the SecurityBundle JUnit report from the green filtered grind.

+++ src/Symfony/Component/Security/Http/EventListener/LoginThrottlingListener.php
@@ -51,12 +51,12 @@ final readonly class LoginThrottlingListener implements EventSubscriberInterface
             // be accepted even if there are 0 tokens remaining to be consumed. We check both
             // anyway for safety in case third party implementations behave unexpectedly.
             if (!$limit->isAccepted() || 0 === $limit->getRemainingTokens()) {
-                throw new TooManyLoginAttemptsAuthenticationException(ceil(($limit->getRetryAfter()->getTimestamp() - time()) / 60));
+                throw new TooManyLoginAttemptsAuthenticationException((int) ceil(($limit->getRetryAfter()->getTimestamp() - time()) / 60));
             }
         } else {
             $limit = $this->limiter->consume($request);
             if (!$limit->isAccepted()) {
-                throw new TooManyLoginAttemptsAuthenticationException(ceil(($limit->getRetryAfter()->getTimestamp() - time()) / 60));
+                throw new TooManyLoginAttemptsAuthenticationException((int) ceil(($limit->getRetryAfter()->getTimestamp() - time()) / 60));
             }
         }
     }
